ci: disable Webpack 5 test in Node v10
Webpack 5 has a transitive dependency on `@npmcli/fs@1.1.0` which doesn't support Node v10, so this test always fails. See: https://app.circleci.com/pipelines/github/angular/angular-cli/19498/workflows/efed3ec6-d3a1-4c2e-b7e6-c2d771d84103/jobs/279530.

+++ tests/legacy-cli/e2e/tests/misc/webpack-5.ts
@@ -4,6 +4,11 @@ import { killAllProcesses, ng, silentYarn } from '../../utils/process';
 import { ngServe, updateJsonFile } from '../../utils/project';
 
 export default async function() {
+  // Webpack 5 has a transitive dependency on `@npmcli/fs@1.1.0`, which doesn't support Node v10.
+  if (process.version.startsWith('v10')) {
+    return;
+  }
+
   // Setup project for yarn usage
   await rimraf('node_modules');
   await updateJsonFile('package.json', (json) => {
